fix: resolve clippy warnings in codirigent-core
- Remove unnecessary u32 cast in session_advanced.rs:637
- Derive Default for TaskManagerConfig instead of manual impl
- Refactor NotesGenerator::generate() to use GenerateNoteParams struct
  - Reduces function arguments from 8 to 2 (self + params)
  - Improves code clarity and maintainability
  - Updates all implementations and tests

+++ crates/codirigent-core/src/session_advanced.rs
@@ -634,7 +634,7 @@ impl OvernightConfig {
 
         let now = chrono::Local::now();
         let hour = now.hour() as u8;
-        let minute = now.minute() as u32;
+        let minute = now.minute();
 
         let end_minutes = (self.stop_hour as u32) * 60;
         let current_minutes = (hour as u32) * 60 + minute;
